% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-interpolate_pca.R
\name{spatial_subset_plot}
\alias{spatial_subset_plot}
\title{Functions for spatial interpolation of PCA}
\usage{
spatial_subset_plot(subset, PC, ...)
}
\description{
Functions for spatial interpolation of PCA
}

