% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTemplate.R
\docType{data}
\name{VizBoundaryTemplate}
\alias{VizBoundaryTemplate}
\title{Data structure handling boundary polygon of a region
using generic input name}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizBoundaryTemplate
}
\description{
Data structure handling boundary polygon of a region
using generic input name
}
\keyword{datasets}

