% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-vizDataSet.R
\docType{data}
\name{VizDataFile}
\alias{VizDataFile}
\title{Generic Data set loaded from file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizDataFile
}
\description{
Data Set with handlers for loading file
}
\keyword{datasets}

