% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizDataSet.R
\docType{data}
\name{VizDataSet}
\alias{VizDataSet}
\title{Generic Data Set}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizDataSet
}
\description{
represents a single data set, which we can think of the
basis unit of analysis.
}
\details{
test
}
\keyword{datasets}

