% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTemplate.R
\docType{data}
\name{VizDataTemplate}
\alias{VizDataTemplate}
\title{Template to load a set of files with predefined names}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizDataTemplate
}
\section{Fields}{

\describe{
\item{\code{templates}}{named vector of strings, constituting default values
to be loaded.}
}}
\keyword{datasets}

