% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizFigure.R
\docType{data}
\name{VizFigure}
\alias{VizFigure}
\title{Basic Figure Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizFigure
}
\description{
VizFigure defines the main module of the Viz App. You wll
likely want to inherit from it for your own figure and datasets. It defines
help, options and all other settings (but most can be overridden when
object is initialized)
}
\section{Fields}{

\describe{
\item{\code{menu}}{menu / analysis this should be displayed in}

\item{\code{id}}{unique, html-nice id}

\item{\code{name}}{human readable name}

\item{\code{size}}{size (in gridster coordinates, i.e. 140 pixel units)}

\item{\code{pos}}{default position (in gridster tiles, from topleft),}

\item{\code{description}}{Brief description of Figure, displayed on questionmark}

\item{\code{info}}{Information about citation, etc, displayed on info sign}

\item{\code{active_on_startup}}{should this be displayed on startup?}

\item{\code{data_sets}}{list of data sets this figures will look for, they should
be of type VizDataSet}

\item{\code{recipe}}{Shiny Module that returns stuff, and creates all the figures}

\item{\code{initialize}}{constructor function}

\item{\code{body}}{Shiny UI describing content of main box.}

\item{\code{get_menu}}{Menu button}

\item{\code{options_sidebar}}{Shiny UI function to display options in right sidebar}
}}
\keyword{datasets}

