% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-vizDataSet.R
\docType{data}
\name{VizMetaData}
\alias{VizMetaData}
\title{Data structure handling meta data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizMetaData
}
\description{
Data structure handling meta data
}
\section{Fields}{

\describe{
\item{\code{files}}{If length one, name of all files, otherwise names of
individual files}
}}
\keyword{datasets}

