% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-vizDataSet.R
\docType{data}
\name{VizPCAData}
\alias{VizPCAData}
\title{Data structure handling pca results data from flashpca}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizPCAData
}
\description{
Data structure handling pca results data from flashpca
}
\section{Fields}{

\describe{
\item{\code{files}}{If length one, name of all files, otherwise names of
individual files}
}}
\keyword{datasets}

