% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTemplate.R
\docType{data}
\name{VizPCATemplate}
\alias{VizPCATemplate}
\title{Data structure handling pca results data from flashpca,
using generic input name}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizPCATemplate
}
\description{
Data structure handling pca results data from flashpca,
using generic input name
}
\keyword{datasets}

