% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizport.R
\docType{data}
\name{VizPort}
\alias{VizPort}
\title{Main class describing viz analysis.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VizPort
}
\description{
not used.
}
\keyword{datasets}

